requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
cffi==1.17.0
    # via cryptography
click==8.1.7
    # via typer
cosl==0.0.26
    # via -r requirements.in
cryptography==43.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via lightkube
idna==3.8
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
lightkube==0.15.4
    # via
    #   -r requirements.in
    #   cosl
lightkube-models==1.30.0.8
    # via
    #   -r requirements.in
    #   lightkube
markupsafe==2.1.5
    # via jinja2
ops==2.16.0
    # via
    #   -r requirements.in
    #   cosl
    #   ops-scenario
ops-scenario==7.0.0
    # via pytest-interface-tester
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   cosl
    #   pytest-interface-tester
pydantic-core==2.20.1
    # via pydantic
pytest==8.3.2
    # via pytest-interface-tester
pytest-interface-tester==3.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   cosl
    #   lightkube
    #   ops
    #   ops-scenario
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.1
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==9.0.0
    # via cosl
typer==0.7.0
    # via pytest-interface-tester
typing-extensions==4.12.2
    # via
    #   cosl
    #   pydantic
    #   pydantic-core
websocket-client==1.8.0
    # via ops
